Change the layout to have a survey file, mostly for readability
