+++ SimPEG/FLOW/Richards/__init__.py
@@ -1,2 +1,3 @@
 from SimPEG.FLOW.Richards import Empirical
+from SimPEG.FLOW.Richards.RichardsSurvey import *
 from SimPEG.FLOW.Richards.RichardsProblem import *
